Extract submission processing logic to separate module
